am62x/am62ax/am62px: sdl: Fix DCC API for Intr Status Checking
SDL_DCC_getStatus was incorrectly checking the whether the Done
and Error Interrupts were enabled in order return the interrupt
status instead of the interrupt status register.

This patch updates SDL_DCC_getStatus to use the status register
to determine whether or not there is a Done or Error interrupt
pending.

Fixes: PROC_SDL-7048

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>