am62x: device_manager: Add dynamic analysis test
The purpose of this testing application is to
achieve the max dynamic analysis code coverage
and branch/decision coverage possible on AM62x

This application can be executed on it's own,
or alongside LDRA for generating the LDRA report.

The scope of this test application is to test
all executable code within device manager's
sciserver and sciclient direct for AM62x.

Fixes: SITSW-3676

Signed-off-by: Rajesh Chinnasamy <Rajesh.Chinnasamy@ti.com>