am62ax/am62x: OSPI: Revert default skip tuning
Currently, OSPI NAND layer enables PHY in the tuning layer once and
performs all reads unlike the OSPI NOR layer which enables PHY, reads
data and then disables it for every read.
This implementation affects the architecture of skip tuning which assumes
the NOR layer of functioning. If skip tuning is enabled on NAND devices,
there is a 50% reduction in performance. A later commit will update the NAND
functionality and it will be similar to the NOR layer where tuning skip
works fine.
This commit only removes the default offering of tuning skip for NAND
devices.

Fixes: SITSW-4229

Signed-off-by: Aashvij Shenai <a-shenai@ti.com>