allsocs: self_reset: Update the host ID used during self reset
ATF sends a message that needs to be forwarded to DM,
this message gets forwarded to DM and disables
interrupt for TIFS incoming thread . But in the
meantime, DM self reset sends messages to TIFS using
the same queue. Since the interrupt is disabled,
TIFS will not be able to respond to the queue. This
creates a race condition in DM/TIFS during boot.
In order to avoid this, use the separate DM2TIFS
forwarding queue for DM self reset. Modify the thread
ID and corresponding host ID of message
accordingly. This race condition is observed in SBL
flow when trace is enabled.

Fixes: SITSW-4243

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>