am62x/am62ax/am62px: ipc: Suspend/Shutdown based on next host state
Use the next host state to decide what to do during suspend, whether
MCU should stay on or be shutdown.

To achieve this the uart wakeup task was redesigned to be a general
suspend task. In this task the next host state is fetched. Depending
on the state the MCU is shutdown or kept running. In case of shutdown
the implementation calls Sciclient_pmSetModuleState() itself.

Fixes: SITSW-4730

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>