am62p: DSS: Enable firewall for DSS register regions used by DM.
- Prepare a list of DSS register regions using sysconfig generated code.
- Configure a background and foreground firewall regions  for DSS
register regions used by DM to enable all access for DM R5 and only read
access for A53.
- APIs to configure firewall are added as part of DSS display share
example and does not imply for other examples and DSS is not shared in
those contexts.
- Firewall for frame buffer region is configured and APIs is not used
as Linux crashes as soon as firewall is configured. -> tracked as part
of separate bug for Linux.

Fixes: SITSW-4731

Signed-off-by: Soumya Tripathy <s-tripathy@ti.com>