am62x: atf: Update the makefiles and linkers for ATF boot.
Build the ATF bl31.bin using the makefiles.
Create the outfile and rprc using the makefile_executable.
Set the boot core ID for the ATF outfile as Load and run.
Set the boot core ID for the boot image as Load only.
Create a appimage containing both the ATF image and the boot image.

Modify the linkers to start from 0x80080000 as initial part of the DDR
memory will be used by ATF.
Modify the linkers to add a new section to ensure the startup code begins
from the ATF jump address.

Fixes: SITSW-5403

Signed-off-by: Yash Srivastava <y-srivastava@ti.com>