allsocs: device_manager: Fix double increment of sequence ID
Currently sequence ID in gSciclientHandle is increment in
the Sciclient_servicePrepareHeader function. This function
is called twice, once in Sciclient_service and again in
Sciclient_serviceSecureProxy. This is causing currSeqId to
increment twice.

Fix this by moving the increment logic to Sciclient_service
from PrepareHeader function as it is called before
Sciclient_serviceSecureProxy is called.

Fixes: SITSW-6190

Signed-off-by: Sparsh Kumar <sparsh-kumar@ti.com>