am62lx: scmi: Enable get parent and set parent functionality
- Add get parent and set parent functionality to soc.c layer when
setting the module clock.
- The commit implements setting of clock based on possible parents
list. Run through the list of parents and set a clock and check whether
the set clock is in a tolerance level of +-5 of required value.
- Add test for set parent, get parent and get possible parents APIs.
- There is a hack fix to set the size of get possible parents payload
to a small value which is lesser than shared memory size. This is in
accordance to TF-A which has a alocation of smaller shared memeory size.

Fixes: SITSW-6545

Signed-off-by: Soumya Tripathy <s-tripathy@ti.com>