am62/am62a/am62p: test: Add tests for lpm abort api
Add the lpm abort api tests in sciclient direct file to increase the
coverage of device manager tests.
These tests will send lpm abort message from DM R5 so as to test the
pass and fail scenario i.e. when message is acked and nacked so that
every portion of lpm abort switch statement is hit.

Fixes: SITSW-4691

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>