j722s: MCU Linker: Support Global Scratch Buffer section similar to WKUP
(Co-Pilot Generated Description below from AI)
This change updates the linker command files for the MCU R5F core on the
J722S to conditionally include a section for a global scratch buffer.
This is similar to the existing WKUP R5F core linker command file, which
already has this section defined.

Fixes: PDK-16822

Signed-off-by: Harshit Gupta <harshit-gupta@ti.com>