am62l: udma: Modify pairing requests
- When we request for autopairing and poll the
paircomplete bit continuously until pairing is
successful, the pairing is not complete sometimes
and the software fails.

- This was observed in stress tests so the auto
pairing requests were removed from being called
before the channel configuration such as priority,
burst size, pdma params is fixed.

- The pairing requests were enabled just before
the channel is enabled which happens only when
all the configuration is fixed.

Fixes: SITSW-7337

Signed-off-by: Nithyaa shri R B <n-rb@ti.com>