am62x/am62ax/am62px: test: SCI: Update the DM Version tests
In order to cover all parts of our DM Version
code, our dynamic analysis tests needed to be
updated accordingly. Both for device manager
and sciclient. Part of this update calls for
changing the order of the DM Version calls
(which are located inside of
sciclient_direct_wrapper.c) which increases
the code coverage

Fixes: SITSW-7239

Signed-off-by: Dillon Frank <d-frank@ti.com>