am62px: dss: non misra C issue fix
ABV.STACK

Buffer overflow-local array index out of bounds
ABV.STACK is an inter-procedural checker that looks for buffer
overflow inside a called function. Arguments passed to the function
control which elements of the buffer are accessed. This checker finds
defects when the buffer is not passed to the function, but is usually
a local variable inside the function or a global variable.

ABV.GENERAL

Buffer overflow-array index out of bounds
A buffer overflow, or overrun, is an anomaly in which a program
writing data to a buffer overruns the buffer's boundaries and
overwrites adjacent memory. Typically, this problem occurs when a
program is copying strings of characters to a buffer.

Fixes: SITSW-6835

Signed-off-by: srirama <srirama@ti.com>