allsocs: mmcsd: Add descriptions for MMCSD module in Sysconfig
The MMCSD module in the Sysconfig allows user to configure different
parameters. In case of no descriptions for the parameters, it is
difficult to understand the usage of the parameters.

This commit adds the descriptions for the required parameters so as to
make it easier for the user to understand the usage and accordingly
configure the parameter.

Additionally, convert the tabs to spaces for consistent rendering of the
lines in the editors.

Fixes: SITSW-7799

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>