am62l: udma: Fix event ISR function to service multiple flow interrupts.
- Previously the flow index updated in chattribute
structure is used to calculate the ringnum for
servicing the registered interrupt.

- This method is ideal for single flow scenario
where the chattribute structure as part of
chhandle is updated only once for one flow.

- But in case of single channel associated with
multiple flows, this attribute gets updated with
latest flowidx configured, so always the latest
flow interrupt will be serviced irrespective of
interrupt occurence.

- So whenever an event is configured for a flow,
the corresponding flow index is updated in
eventhandle->flownum and passed to ISR function as
argument. In this way the proper handle with
proper flownum is serviced.

Fixes: SITSW-7436

Signed-off-by: Nithyaa shri R B <n-rb@ti.com>