am62a/am62p: tools: generate der cert and update copyright year
The debug unlock certificate generation script generates the certificate
in the hex format only discarding the DER certificate blob. This works
only for those who are using the TISCI message for JTAG unlocking.

The TIFS supports JTAG unlocking via the CCS as well in which case the
DER certificate blob is required.

- Fix the script to generate the DER certificate blob as as well.
- Fix the copyright year as well by dynamically deriving it using the
datetime library.

Fixes: SITSW-7832, SITSW-7833

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>