all: udma: Add UART DMA multi-instance example
- Add examples to demonstrate usage of multiple
udma instances with uart.

- Currently there is example to test one dma
instance configured, but this did not capture
uart template bug which did not pass proper handle
based on the udma instance configured.

- So when configuring multiple dma instants there
was no example to test such scenarios and hence
the template bug could not be captured.

- This template bug was fixed to assign proper dma
handles based on the dma instance configured
(config_udma0/config_udma1) and added uart
example with two dma instances to test the same.

Fixes: SITSW-7922, SITSW-8008

Signed-off-by: Nithyaa shri R B <n-rb@ti.com>