all: mcspi: Update force bit properly for DMA mode of operation
- When cs disable is false the force field in
channel config register is not set properly.

- There is a chconfigregval member in the chobj
structure which holds mcspi chconfig register
value.

- The force bit which is one of the field in
chconfig register is set based on csenable is
true/false inside the driver.

- So when the force bit is asserted the chconfig
register value needs to be updated in the
chconfigregval member, but this is missed in DMA
flow of operation and not in other modes.

- So when application enters mcspi-chsetdatasize
API the same chnconfig register is updated for
datasize using the chconfigregval member,
as the force bit status is not updated
properly in chconfigregval (force = 0 as per this
member but force=1 in the register space).
This Force bit gets reset though it is enabled in
mcspi register, which stalls the transaction.

- Update the chnconfigregval member whenever the
chnconfig register is modified, so that the
structure member and the actual register values
are in sync.

Fixes: SITSW-7171, SITSW-8055, SITSW-8056,
       SITSW-8057

Signed-off-by: Nithyaa shri R B <n-rb@ti.com>