allsocs: boot: Fix the gAppimage section allocation
The `gAppimage` variable is meant to be allocated to the `.bss.app`
section as defined in the linker command files. However, in some SBLs,
it is allocated to the `.app` section which is different than the
`.bss.app` section.

Fix the allocation by explicitly allocating the variable to `.bss.app`
section.

Fixes: SITSW-8456

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>