am62x/am62ax/am62px/am62dx: boot: fix dma disabled for soc memory
The bootloader driver is not using DMA for parsing and loading the
appimage for the memory bootloader instances.

This is because the bootloader driver uses the given memory bootloader
instance only for parsing and loading the appimage instead of using the
global memory bootloader instance which already has the DMA enabled and
is used for other boot media like NOR or eMMC.

Fix the issue by using the global memory bootloader instance for all
boot media to make sure the DMA is used for parsing and loading of the
appimage thus significantly improving the boot time.

Fixes: SITSW-8380

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>