all: i2c: Handle TX underflow in target mode
When the master requests more bytes than the slave can provide,
the transfer gets stuck waiting on the completion semaphore
because the slave has no data left to send. To prevent the hang,
we now feed a dummy byte 0xFFU whenever a TX underrun occurs,
allowing the transfer to continue and complete.

Fixes: SITSW-8272

Signed-off-by: Piyush Sharma <p-sharma3@ti.com>