all: ospi: move NOR flash boot fixup to driver
With some BOOTMODE configurations, the ROM code might switch to a higher
speed mode (like 8D-8D-8D) after reading SFDP. However, our driver
expects the Flash to be in 1S-1S-1S mode at the time of initialization.
Hence, add the fixup that switches it back to 1S-1S-1 in the driver
code.

With this change the manual fixups in examples and tests are no longer
required.

Fixes: SITSW-8937

Signed-off-by: Amneesh Singh <amneesh@ti.com>