am62d: test: Fix ddr_copy test
The ddr copy test check if ddr copy is performed when in place
authentication fails. The test writes a corrupted image to the flash and
when it is loaded, the authentication fails. The image used for testing
on am62d hsse is not corrupted, updated the image with corrupt image.

Fixes: SITSW-9017

Signed-off-by: Nikhil M Jain <n-jain1@ti.com>