allsocs: dm: Fix the NACK response handling
The NACK response handling is currently broken for LPM messages that are
processed in sciclient_service function. Also, in case of query fw caps
api and RM board config, the handling of NACK case is improper.
Fix that by setting response to NACK in case of failure. Also, return
CSL_PASS if DM has handled the message and set NACK for any incorrect
parameters or TIFS NACK.

Fixes: SITSW-9038

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>