all: lpm: Add WFI in MCU only mode to save power
Currently, in the LPM examples, the MCU core waits to receive an
interrupt from DMTimer or MCAN or MCU UART or IPC message to exit from
MCU Only mode.
This wait is a polling loop due to which the core consumes a lot of
power. Change it to WFI so that the power can be optimized.

Fixes: SITSW-9713

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>